refactor(input): split input.rs into per-context handler modules
Move-only split of input.rs (~1470 lines), no behaviour change:

- mod.rs — handle_event: overlay dispatch + the global key match
  (561 lines; the match itself is the irreducible part).
- geometry.rs — neighbors_d1, face_d2, list_nav and their tests.
- One file per wizard handler: travel, repair, mine, jettison, map,
  craft (manny + atomic printer), pickers (salvage, recall, rename,
  deploy, inspect, recover, detach — mirrors ui/overlays/pickers.rs),
  scanner (waypoints, object actions + dispatch).

Handlers become pub(super); the public surface of the module is
unchanged (input::handle_event).
