+++ src/input/travel.rs
@@ -1,0 +1,35 @@
+use crossterm::event::KeyCode;
+use tokio::sync::mpsc;
+
+use crate::api::client::ApiClient;
+use crate::api::tasks::fetch_move;
+use crate::app::{
+    ApiMessage, AppState, TravelInput,
+};
+pub(super) fn handle_travel_event(
+    code: KeyCode,
+    state: &mut AppState,
+    client: &ApiClient,
+    tx: &mpsc::Sender<ApiMessage>,
+) {
+    match &state.travel {
+        TravelInput::Typing(_) => match code {
+            KeyCode::Esc => state.travel = TravelInput::Inactive,
+            KeyCode::Backspace => state.travel_backspace(),
+            KeyCode::Enter => state.travel_submit(),
+            KeyCode::Char(c) => state.travel_type_char(c),
+            _ => {}
+        },
+        TravelInput::Confirming { x, y, z, error, .. } => {
+            let (x, y, z, has_error) = (*x, *y, *z, error.is_some());
+            match code {
+                KeyCode::Esc => state.travel = TravelInput::Inactive,
+                KeyCode::Enter if !has_error => {
+                    fetch_move(x, y, z, client.clone(), tx.clone());
+                }
+                _ => {}
+            }
+        }
+        TravelInput::Inactive => {}
+    }
+}
